download images from csv

+++ src/utils/download_images_from_csv.py
@@ -1,4 +1,5 @@
 
+from fileinput import filename
 import sys
 
 
@@ -16,10 +17,17 @@ from joblib import Parallel, delayed
 from active_fire.general import ActiveFireIndex
 from image.sentinel import BufferedImageStack
 from image.converter import convert_dir_jp2_to_tiff, get_cloud_mask
+from utils.metadata import get_image_metadata
+import time
+from tqdm import tqdm
+import csv
 
-N_JOBS = 2
+N_JOBS = -2
+# N_JOBS = 3
+
 
 INPUT_CSV = '../../resources/s2_continent_202008_list_to_download.csv'
+OUTPUT_CSV = '../../resources/downloaded_images.csv'
 
 DOWNLOAD_PATH = '../../resources/images/download/'
 OUTPUT_PATH = '../../resources/images/original/'
@@ -36,46 +44,100 @@ NON_CLASSIFICATION_BANDS = tuple(set(SENTINEL_BANDS) - set(CLASSIFICATION_BANDS)
 
 ACTIVE_FIRE_ALGORITHMS = [
     {'method': 'Liangrocapart'},
+    {'method': 'KatoNakamura'},
+    {'method': 'Yongxue'},
+    {'method': 'Murphy'},
 ]
 
 
 FIRE_GRIDS = []
 
-def check_fire_in_tile(tiff_path, cloud_path=None):
-    try:
-
-        # Load the bands to a buffer
-        bands_files = glob(os.path.join(tiff_path, '*.tif'))
-        img_buffer = BufferedImageStack()
-        for band_file in bands_files:
-            band = 12
-            if band_file.endswith('_B11.tif'):
-                band = 11
-            elif band_file.endswith('_B8A.tif'):
-                band = '8A'
-            
-            img_buffer.load_file_as_band(band_file, band)
-
-        image_shape = img_buffer.read(12).shape
-        cloud_mask = np.ones(image_shape, dtype=np.bool) 
+def check_fire_in_tile(tiff_path, metadata):
+    # Load the bands to a buffer
+    bands_files = glob(os.path.join(tiff_path, '*.tif'))
+    img_buffer = BufferedImageStack()
+    
+    shape = None, None
+    for band_file in bands_files:
+        band = 12
+        if band_file.endswith('_B11.tif'):
+            band = 11
+        elif band_file.endswith('_B8A.tif'):
+            band = '8A'
         
+        img_buffer.load_file_as_band(band_file, band)
+    
+            
+        if shape[0] is not None and img_buffer.metas[band]['width'] != shape[0] or img_buffer.metas[band]['height'] != shape[1]:
+            # raise Exception("Images with different shapes: {} - {}x{}".format(tiff_path, img_buffer.metas[band]['width'], img_buffer.metas[band]['height']))
+            return None
+        shape = img_buffer.metas[band]['width'], img_buffer.metas[band]['height']
 
-        if cloud_path is not None and os.path.exists(cloud_path):
-            cloud_mask = get_cloud_mask(cloud_path, image_shape, img_buffer.metas[12]['transform'])
+    
+    file_name = os.path.basename(bands_files[0])
+    file_name = file_name.split('_')
+    
+    grid_name = file_name[0]
+    timestamp = file_name[1]
+
+    # Check for fire
+    total_fire_pixels = 0
+
+    # data = []
+    # for algorithm in ACTIVE_FIRE_ALGORITHMS:
+        # afi = ActiveFireIndex(algorithm['method'])
+        # start_time = time.time()
+        # mask = afi.transform(buffered_stack=img_buffer, metadata=metadata)
+        # end_time = time.time()
+
+        # num_fire_pixels = mask.sum()
+        # total_fire_pixels += num_fire_pixels
+
+        # data.append({
+        #     'method' : algorithm['method'],
+        #     'num_fire_pixels': num_fire_pixels,
+        #     'grid_name': grid_name,
+        #     'timestamp' : timestamp,
+        #     'tmp_path' : tiff_path,
+        #     'start_time': start_time,
+        #     'end_time': end_time,
+        #     'delta_time': (end_time - start_time),
+        # })
+        # print('{} - Num. fire pixels: {}'.format(algorithm['method'], num_fire_pixels))
         
-        for algorithm in ACTIVE_FIRE_ALGORITHMS:
-            
-            afi = ActiveFireIndex(algorithm['method'])
-            mask = afi.transform(img_buffer)
-            num_fire_pixels = (mask & cloud_mask).sum()
+    
+    data = Parallel(n_jobs=N_JOBS, verbose=0)(delayed(process_image)(algorithm, img_buffer, metadata) for algorithm in ACTIVE_FIRE_ALGORITHMS)
+    for d in data:
+        d['grid_name'] = grid_name
+        d['timestamp'] = timestamp
+        d['tmp_path'] = tiff_path
 
-            # print('{} - Num. fire pixels: {}'.format(algorithm['method'], num_fire_pixels))
-            
-            return num_fire_pixels
+        total_fire_pixels += d['num_fire_pixels']
 
-    except Exception as e:
-        # Abort in case of error
-        return None
+    with open(OUTPUT_CSV, 'a+') as out:
+        writer = csv.DictWriter(out, fieldnames=['method', 'num_fire_pixels', 'grid_name', 'timestamp', 'tmp_path', 'delta_time'])
+        writer.writerows(data)    
+    
+    return total_fire_pixels
+
+def process_image(algorithm, img_buffer, metadata):
+    afi = ActiveFireIndex(algorithm['method'])
+    start_time = time.time()
+    mask = afi.transform(buffered_stack=img_buffer, metadata=metadata)
+    end_time = time.time()
+
+    num_fire_pixels = mask.sum()
+    
+    return {
+        'method' : algorithm['method'],
+        'num_fire_pixels': num_fire_pixels,
+        # 'grid_name': grid_name,
+        # 'timestamp' : timestamp,
+        # 'tmp_path' : tiff_path,
+        # 'start_time': start_time,
+        # 'end_time': end_time,
+        'delta_time': (end_time - start_time),
+    }
 
 
 def download_mask_cloud(file):
@@ -96,7 +158,7 @@ def download_mask_cloud(file):
         return cloud_file
 
     except Exception as e:
-        log_file = os.path.join(LOG_PATH, '{}_MSK_CLOUDS_B00.txt'.format(tile_name))
+        log_file = os.path.join(LOG_PATH, '{}_MSK_CLOUDS_B00.txt.error'.format(tile_name))
         with open(log_file, 'w+') as f:
             f.write(str(e))
 
@@ -114,15 +176,14 @@ def download_metadata(file):
         response = requests.get(metadata_url)
         metadata_folder = os.path.join(OUTPUT_METADATA_PATH, tile_name)
         os.makedirs(metadata_folder, exist_ok=True)
-
         mtd_tl = os.path.join(metadata_folder, 'MTD_TL.xml') 
         with open(mtd_tl, 'wb') as f:
             f.write(response.content)
     
-        metadata_url = os.path.join(file.split('GRANULE')[0], tile_name+'_MTD_MSIL1C.xml') 
-
+        metadata_url = os.path.join(file.split('GRANULE')[0], 'MTD_MSIL1C.xml') 
         response = requests.get(metadata_url)
-        mtd_msi = os.path.join(metadata_folder, tile_name+'_MTD_MSIL1C.xml') 
+        
+        mtd_msi = os.path.join(metadata_folder, 'MTD_MSIL1C.xml') 
         with open(mtd_msi, 'wb') as f:
             f.write(response.content)
 
@@ -130,7 +191,7 @@ def download_metadata(file):
         return mtd_tl, mtd_msi
 
     except Exception as e:
-        log_file = os.path.join(LOG_PATH, '{}_MTD.txt'.format(tile_name))
+        log_file = os.path.join(LOG_PATH, '{}_MTD.txt.error'.format(tile_name))
         with open(log_file, 'w+') as f:
             f.write(str(e))
 
@@ -141,35 +202,42 @@ def download_metadata(file):
 
 def download_sentinel_bands(file, bands_to_download):
     tile_name = os.path.basename(file)
-    
     try:     
         download_path = os.path.join(DOWNLOAD_PATH, tile_name)
-        for band in bands_to_download:
-            band_url = '{}_{}.jp2'.format(file, band)
-            file_name = os.path.basename(band_url)
-            log_file = os.path.join(LOG_PATH, file_name.replace('.jp2', '.txt'))
 
-            if os.path.exists(log_file):
-                continue
-            
-            response = requests.get(band_url)
-            output_file = os.path.join(download_path, file_name)
-            with open(output_file, 'wb') as f:
-                f.write(response.content)
-            
-            with open(log_file, 'w+') as f:
-                f.write('200\n')
-
+        # for band in bands_to_download:
+        #     download_sentinel_band(file, band, download_path)
         
+        Parallel(n_jobs=N_JOBS, verbose=0)(delayed(download_sentinel_band)(file, band, download_path) for band in bands_to_download)
+
         return download_path
 
     except Exception as e:
-        log_file = os.path.join(LOG_PATH, '{}.log'.format(tile_name))
+        log_file = os.path.join(LOG_PATH, '{}.log.error'.format(tile_name))
         with open(log_file, 'w+') as f:
             f.write(str(e))
 
-    return None, None
+    return None
 
+
+def download_sentinel_band(file, band, download_path):
+    band_url = '{}_{}.jp2'.format(file, band)
+    file_name = os.path.basename(band_url)
+    log_file = os.path.join(LOG_PATH, file_name.replace('.jp2', '.txt'))
+
+    if os.path.exists(log_file):
+        return
+    
+    response = requests.get(band_url)
+    output_file = os.path.join(download_path, file_name)
+    with open(output_file, 'wb') as f:
+        f.write(response.content)
+    
+    with open(log_file, 'w+') as f:
+        f.write('200\n')
+
+
+    return download_path
 
 def download_file(file):
     
@@ -190,18 +258,23 @@ def download_file(file):
            FIRE_GRIDS.append(grid_name)
         return
     
+
     # Ignore if there are another image of the grid with fire
     if grid_name in FIRE_GRIDS:
         with open(log_file, 'w+') as f:
             f.write('TILE-FIRE')
+        
+        with open(OUTPUT_CSV, 'a+') as out:
+            writer = csv.writer(out)
+            writer.writerow(['Ignore', 'tile-file', 'tile found fire before']) 
+                
         return
 
     tmp_dir = os.path.join(DOWNLOAD_PATH, tile_name, 'tmp')
     os.makedirs(tmp_dir, exist_ok=True)
 
-    # Download the cloud mask
-    cloud_file = download_mask_cloud(file)
     download_path = download_sentinel_bands(file, CLASSIFICATION_BANDS)
+
     if download_path is None:
         with open(log_file, 'a+') as f:
             f.write('ERROR')
@@ -211,20 +284,19 @@ def download_file(file):
     if mtd_tl is None or mtd_msi is None:
         with open(log_file, 'a+') as f:
             f.write('ERROR')
-        return 
+        raise Exception('Error with metadata: {}'.format(file))
+        
+    metadata = get_image_metadata(mtd_tl_xml=mtd_tl, mtd_msil_xml=mtd_msi)
 
     convert_dir_jp2_to_tiff(download_path, tmp_dir, verbose=False)
-        
-    num_fire_pixels = check_fire_in_tile(tmp_dir, cloud_path=cloud_file)
+    num_fire_pixels = check_fire_in_tile(tmp_dir, metadata)
 
-    if num_fire_pixels is None:
-        has_fire = False
-    else:
-        has_fire = (num_fire_pixels > 0)
-
-    if has_fire:
+    if num_fire_pixels is not None and num_fire_pixels > 0:
         # Save 
         FIRE_GRIDS.append(grid_name)
+
+        # Download the cloud mask
+        cloud_file = download_mask_cloud(file)
 
         # Download the missing bands
         download_path = download_sentinel_bands(file, NON_CLASSIFICATION_BANDS)
@@ -232,7 +304,7 @@ def download_file(file):
         if download_path is None:
             with open(log_file, 'a+') as f:
                 f.write('ERROR')
-            return
+            raise Exception('Error downloading: {}'.format(download_path))
 
         # Convert the bands to TIFF
         convert_dir_jp2_to_tiff(download_path, tmp_dir, verbose=False)
@@ -250,25 +322,25 @@ def download_file(file):
 
     else:
         # Remove the cloud mask file
-        if os.path.exists(cloud_file):
-            os.remove(cloud_file)
+        # if cloud_file is not None and os.path.exists(cloud_file):
+        #     os.remove(cloud_file)
 
-        if os.path.exists(mtd_tl):
+        if mtd_tl is not None and os.path.exists(mtd_tl):
             os.remove(mtd_tl)
 
-        if os.path.exists(mtd_msi):
+        if mtd_msi is not None and os.path.exists(mtd_msi):
             os.remove(mtd_msi)
 
         # Register that the image was processed
         with open(log_file, 'w+') as f:
-            if has_fire is None:
+            if num_fire_pixels is None:
                 f.write('ERROR')
             else:
                 f.write('NO-FIRE')
 
     # Delete the temporary files
     try:
-        if os.path.exists(download_path):
+        if download_path is not None and os.path.exists(download_path):
             shutil.rmtree(download_path)
     except:
         pass
@@ -281,12 +353,25 @@ if __name__ == '__main__':
     os.makedirs(LOG_PATH, exist_ok=True)
 
     df = pd.read_csv(INPUT_CSV,  sep=';')
-    files = df.url.unique()
-    files = [f for f in files if str(f) != 'nan']
-    files.sort()
 
 
-    # for ind, file in enumerate(files[:1]):
+    # files = df.url.unique()
+    # files = [f for f in files if str(f) != 'nan']
+    # files.sort()
+    # for ind, file in enumerate(files):
     #     download_file(file)
         
-    Parallel(n_jobs=N_JOBS, verbose=100)(delayed(download_file)(file) for file in files)
+    # Group by tile to avoid download two times the same spot (if fire were found)
+    g = df.groupby('sp1').size()
+    for tile, count in tqdm(g.items(), total=len(g)):
+        df_tile = df[ df['sp1'] == tile ]
+
+        files = df_tile.url.unique()
+        files = [f for f in files if str(f) != 'nan']
+        files.sort()
+        
+
+        for file in files:
+            download_file(file)
+            
+        # Parallel(n_jobs=N_JOBS, verbose=100)(delayed(download_file)(file) for file in files)
